+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 * Place wrappers in right order to not lose any data after applying `flow()`. (UIDATIMP-1716)
 * Fix stuck Progress Bar in import files. (UIDATIMP-1725)
 * Fix the 'Existing record field' list does not contain criterion value types in the match profile. (UIDATIMP-1736)
+* Match profiles details view: incoming -> existing record table view breaks when panel is narrowed (UIDATIMP-1751)
 
 ## [9.0.0](https://github.com/folio-org/ui-data-import/tree/v9.0.0) (2025-03-13)